fix(ios): stop the accueil ruling on a month it has not seen
Until something is pointed the estimate equals the plan by construction, so
the hero told a brand-new account it was conforme à son budget, under a chart
still saying it waits for a first pointing. The verdict now reads the same
signal the plot does. Its neighbour metric is a count of operations, so the
variance beside it carries its currency rather than passing for one too.

Both sentences move onto PresentationState, next to the accessibility one
that duplicated the same three-way branch — the shape that produced the
DriftCard contradiction. The trajectory tests leave with the 500-line ceiling,
into the formulas folder where what they exercise already lives.
